Create a dataframe object from CES survey.

get_ces() creates a dataframe for a requested Canadian Election Study survey using an associated survey code to call and download the survey dataset. Prints out the associated citation for use with the requested dataset and a link to the orginal location of the data file.

get_ces(srvy)

Arguments

srvy
A CES survey code call. See Survey Code Calls below. srvy value must be a character string.

Details

Datasets

Datasets are loaded using either .dta or .sav file types. To quickly convert a dataset's values to factor type use labelled::to_factor() on the dataset.

Survey Code Calls

ces2019_web calls 2019 CES online survey dataset.

ces2019_phone calls 2019 CES phone survey dataset.

ces2015_web calls 2015 CES online survey dataset.

ces2015_phone calls 2015 CES phone survey dataset.

ces2015_combo calls 2015 CES combined (online/phone) dataset.

ces2011 calls 2011 CES survey dataset.

ces2008 calls 2008 CES survey dataset.

ces2004 calls 2004 CES survey dataset.

ces0411 calls 2004-2011 CES survey dataset.

ces0406 calls 2004-2006 CES survey dataset.

ces2000 calls 2000 CES survey dataset.

ces1997 calls 1997 CES survey dataset.

ces1993 calls 1993 CES survey dataset.

ces1988 calls 1988 CES survey dataset.

ces1984 calls 1984 CES survey dataset.

ces1974 calls 1974 CES survey dataset.

ces7480 calls 1974-1980 CES survey dataset.

ces72_jnjl calls 1972 June-July CES survey dataset.

ces72_sep calls 1972 September CES survey dataset.

ces72_nov calls 1972 November CES survey dataset.

ces1968 calls 1968 CES survey dataset.

ces1965 calls 1965 CES survey dataset.

Incorrect/Repeated Code Calls

Incorrect (a non-existent survey code) will stop the function process and return associated error messages. Repeated code calls will load in an unaltered version of the requested table.

Extra Notes

Due to the naming of the columns in the 1965 and 1968 datasets it is recommended to download the associated codebook for the requested dataset.

Examples

devtools::install_github("hodgettsp/cesR")

#> Downloading GitHub repo hodgettsp/cesR@HEAD

#> cpp11 (0.1.0 -> 0.2.1) [CRAN]
#> dplyr (1.0.0 -> 1.0.2) [CRAN]
#> tidyr (1.1.0 -> 1.1.1) [CRAN]

#> Installing 3 packages: cpp11, dplyr, tidyr

#> Installing packages into 'C:/Users/phodg/AppData/Local/Temp/RtmpcJ6NKt/temp_libpath104c3fb57930'
#> (as 'lib' is unspecified)

#> package 'cpp11' successfully unpacked and MD5 sums checked
#> package 'dplyr' successfully unpacked and MD5 sums checked
#> package 'tidyr' successfully unpacked and MD5 sums checked
#> 
#> The downloaded binary packages are in
#> 	C:\Users\phodg\AppData\Local\Temp\RtmpkLTDD7\downloaded_packages
#>          checking for file 'C:\Users\phodg\AppData\Local\Temp\RtmpkLTDD7\remotesfc7e1c489d\hodgettsp-cesR-a8278d5/DESCRIPTION' ...  v  checking for file 'C:\Users\phodg\AppData\Local\Temp\RtmpkLTDD7\remotesfc7e1c489d\hodgettsp-cesR-a8278d5/DESCRIPTION'
#>       -  preparing 'cesR':
#>    checking DESCRIPTION meta-information ...     checking DESCRIPTION meta-information ...   v  checking DESCRIPTION meta-information
#>       -  checking for LF line-endings in source and make files and shell scripts
#>       -  checking for empty or unneeded directories
#>   -  building 'cesR_0.1.0.tar.gz'
#>      
#> 

#> Error: Failed to install 'cesR' from GitHub:
#>   (converted from warning) package 'cesR' is in use and will not be installed


library(cesR)

get_ces("ces2019_phone")

#> TO CITE THIS SURVEY FILE: Stephenson, Laura B; Harell, Allison; Rubenson, Daniel; Loewen, Peter John, 2020, '2019 Canadian Election Study - Phone Survey', https://doi.org/10.7910/DVN/8RHLG1, Harvard Dataverse, V1, UNF:6:eyR28qaoYlHj9qwPWZmmVQ== [fileUNF]
#> 
#> LINK: https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/8RHLG1


ces2019_phone <- labelled::to_factor(ces2019_phone)
head(ces2019_phone)

#> # A tibble: 6 x 278
#>   sample_id survey_end_CES survey_end_mont~ survey_end_day_~ num_attempts_CES
#>       <dbl> <chr>                     <dbl>            <dbl>            <dbl>
#> 1        18 2019-09-23 15~                9               23                5
#> 2        32 2019-09-12 18~                9               12                1
#> 3        39 2019-09-10 18~                9               10                1
#> 4        59 2019-10-10 15~               10               10                6
#> 5        61 2019-09-12 16~                9               12                1
#> 6        69 2019-09-17 17~                9               17                1
#> # ... with 273 more variables: interviewer_id_CES <dbl>,
#> #   interviewer_gender_CES <chr>, language_CES <fct>, phonetype_CES <fct>,
#> #   survey_end_PES <chr>, survey_end_month_PES <fct>, survey_end_day_PES <fct>,
#> #   num_attempts_PES <fct>, interviewer_id_PES <fct>,
#> #   interviewer_gender_PES <chr>, language_PES <fct>, phonetype_PES <fct>,
#> #   mode_PES <fct>, phone_type <fct>, weight_CES <dbl>, weight_PES <fct>,
#> #   c1 <fct>, c2a <fct>, c3 <fct>, q1 <fct>, q2 <fct>, q3 <fct>, q4 <fct>,
#> #   q6 <fct>, q7 <chr>, q72 <chr>, q73 <chr>, q74 <chr>, q8 <fct>, q8_70 <chr>,
#> #   q9 <fct>, q10 <fct>, q11 <fct>, q11_70 <chr>, q12 <fct>, q12_70 <chr>,
#> #   q13 <fct>, q14 <fct>, q15 <fct>, q16 <fct>, q17 <fct>, q18 <fct>,
#> #   q19 <fct>, q20 <fct>, q21 <fct>, q22 <fct>, q23 <fct>, q24 <fct>,
#> #   q25 <fct>, q27_a <fct>, q27_b <fct>, q27_c <fct>, q27_d <fct>, q27_e <fct>,
#> #   q31 <fct>, q32 <fct>, q33 <fct>, q33_70 <chr>, q34 <fct>, q34_70 <chr>,
#> #   q35 <fct>, q35_70 <chr>, q36 <fct>, q36_70 <chr>, q37 <fct>, q37_70 <chr>,
#> #   q38 <fct>, q38_70 <chr>, q39 <fct>, q40 <fct>, q75 <fct>, q44 <fct>,
#> #   q76 <fct>, q45 <fct>, q46 <fct>, q47 <fct>, q48 <fct>, q49 <fct>,
#> #   q52 <fct>, q52_70 <chr>, q53 <fct>, q54 <fct>, q59 <fct>, q60 <fct>,
#> #   q60_70 <chr>, q77 <fct>, q43 <fct>, q61 <fct>, q62 <fct>, q62_220 <chr>,
#> #   q63 <fct>, q64 <fct>, q64_130 <chr>, q65 <fct>, q66a_1 <fct>, q66a_2 <fct>,
#> #   q66a_3 <fct>, q66a_4 <fct>, q66a_5 <fct>, q66a_6 <fct>, ...