Print a dataset of the CES survey codes and their associated calls.

get_cescodes() prints out a data frame of the CES survey codes and the associated calls. Provides a quick way of looking up a CES survey code and the associated calls.

get_cescodes()

Details

Items under the Survey Code Calls and Index Code Calls can be copied and used with the get_ces() function.

Survey Codes

ces2019_web 2019 CES online survey dataset.

ces2019_phone 2019 CES phone survey dataset.

ces2015_web 2015 CES online survey dataset.

ces2015_phone 2015 CES phone survey dataset.

ces2015_combo 2015 CES combined (online/phone) dataset.

ces2011 2011 CES survey dataset.

ces2008 2008 CES survey dataset.

ces2004 2004 CES survey dataset.

ces0411 2004-2011 CES survey dataset.

ces0406 2004-2006 CES survey dataset.

ces2000 2000 CES survey dataset.

ces1997 1997 CES survey dataset.

ces1993 1993 CES survey dataset.

ces1988 1988 CES survey dataset.

ces1984 1984 CES survey dataset.

ces1974 1974 CES survey dataset.

ces7480 1974-1980 CES survey dataset.

ces72_jnjl 1972 June-July CES survey dataset.

ces72_sep 1972 September CES survey dataset.

ces72_nov 1972 November CES survey dataset.

ces1968 1968 CES survey dataset.

ces1965 1965 CES survey dataset.

Survey Code Calls

"ces2019_web" calls 2019 CES online survey dataset.

"ces2019_phone" calls 2019 CES phone survey dataset.

"ces2015_web" calls 2015 CES online survey dataset.

"ces2015_phone" calls 2015 CES phone survey dataset.

"ces2015_combo" calls 2015 CES combined (online/phone) dataset.

"ces2011" calls 2011 CES survey dataset.

"ces2008" calls 2008 CES survey dataset.

"ces2004" calls 2004 CES survey dataset.

"ces0411" calls 2004-2011 CES survey dataset.

"ces0406" calls 2004-2006 CES survey dataset.

"ces2000" calls 2000 CES survey dataset.

"ces1997" calls 1997 CES survey dataset.

"ces1993" calls 1993 CES survey dataset.

"ces1988" calls 1988 CES survey dataset.

"ces1984" calls 1984 CES survey dataset.

"ces1974" calls 1974 CES survey dataset.

"ces7480" calls 1974-1980 CES survey dataset.

"ces72_jnjl" calls 1972 June-July CES survey dataset.

"ces72_sep" calls 1972 September CES survey dataset.

"ces72_nov" calls 1972 November CES survey dataset.

"ces1968" calls 1968 CES survey dataset.

"ces1965" calls 1965 CES survey dataset.

See also

Examples

devtools::install_github("hodgettsp/cesR")

#> Downloading GitHub repo hodgettsp/cesR@HEAD

#> 
#>          checking for file 'C:\Users\phodg\AppData\Local\Temp\RtmpCCHQaj\remotes22384d0b35c8\hodgettsp-cesR-0212deb/DESCRIPTION' ...     checking for file 'C:\Users\phodg\AppData\Local\Temp\RtmpCCHQaj\remotes22384d0b35c8\hodgettsp-cesR-0212deb/DESCRIPTION' ...   v  checking for file 'C:\Users\phodg\AppData\Local\Temp\RtmpCCHQaj\remotes22384d0b35c8\hodgettsp-cesR-0212deb/DESCRIPTION' (720ms)
#>       -  preparing 'cesR': (440ms)
#>    checking DESCRIPTION meta-information ...     checking DESCRIPTION meta-information ...   v  checking DESCRIPTION meta-information
#>       -  checking for LF line-endings in source and make files and shell scripts
#>       -  checking for empty or unneeded directories
#>       -  building 'cesR_0.1.0.tar.gz'
#>      
#> 

#> Error: Failed to install 'cesR' from GitHub:
#>   (converted from warning) package 'cesR' is in use and will not be installed


library(cesR)

get_cescodes()

#>    index ces_survey_code get_ces_call_char
#> 1      1     ces2019_web     "ces2019_web"
#> 2      2   ces2019_phone   "ces2019_phone"
#> 3      3     ces2015_web     "ces2015_web"
#> 4      4   ces2015_phone   "ces2015_phone"
#> 5      5   ces2015_combo   "ces2015_combo"
#> 6      6         ces2011         "ces2011"
#> 7      7         ces2008         "ces2008"
#> 8      8         ces2004         "ces2004"
#> 9      9         ces0411         "ces0411"
#> 10    10         ces0406         "ces0406"
#> 11    11         ces2000         "ces2000"
#> 12    12         ces1997         "ces1997"
#> 13    13         ces1993         "ces1993"
#> 14    14         ces1988         "ces1988"
#> 15    15         ces1984         "ces1984"
#> 16    16         ces1974         "ces1974"
#> 17    17         ces7480         "ces7480"
#> 18    18      ces72_jnjl      "ces72_jnjl"
#> 19    19       ces72_sep       "ces72_sep"
#> 20    20       ces72_nov       "ces72_nov"
#> 21    21         ces1968         "ces1968"
#> 22    22         ces1965         "ces1965"