Creates a non-exhaustive dataframe of 21 variables with renamed columns.

get_decon() creates a non-exhaustive dataset under the name decon consisting of 21 variables with renamed columns from the demographics, ideology, and economy sections of the 2019 CES online survey.

get_decon(pos = 1)

Arguments

pos
Environment assignment. Defaults to 1, which is an assignment to the global environment.

decon Variables

citizenship

Canadian citizenship status

yob

year of birth

gender

identified gender of the respondent

province_territory

Province or Territory of current residence

education

highest level of education completed

vote_likely

likelihood of voting on election day

vote_likely_ifable

likelihood to vote in first election for which respondent is eligible

votechoice

party most likely to vote for

votechoice_text

party most likely to vote for - text answers

votechoice_couldvote

party most likely to vote for if eligible to vote

votechoice_couldvote_text

party most likely to vote for if eligible to vote - text answers

votechoice_unlikely

party least likely to vote for

votechoice_unlikely_text

party least likely to vote for - text answers

votechoice_unlikely_couldvote

party least likely to vote for if eligible to vote

votechoice_unlikely_couldvote_text

party least likely to vote for if eligible to vote - text answers

vote_advancevote_choice

party voted for in the advanced ballot

vote_advancevote_choice_text

party voted for in the advanced ballot - text

vote_partylean

party toward which the respondent leans

vote_partylean_text

party toward which the respondent leans - text answers

vote_partylean_couldvote

party toward which the respondent leans if eligible

vote_partylean_couldvote_text

party toward which the respondent leans if eligible - text answers

votechoice_secondchoice

second choice party of respondent

votechoice_secondchoice_text

second choice party of respondent - text answers

votechoice_couldvote_secondchoice

second choice party of respondent if eligible

votechoice_couldvote_secondchoice_text

second choice party of respondent if eligible - text answers

votechoice_partynotvote_1

party respondent would note vote for - first ranking

votechoice_partynotvote_2

party respondent would note vote for - second ranking

votechoice_partynotvote_3

party respondent would note vote for - third ranking

votechoice_partynotvote_4

party respondent would note vote for - fourth ranking

votechoice_partynotvote_5

party respondent would note vote for - fifth ranking

votechoice_partynotvote_6

party respondent would note vote for - sixth ranking

votechoice_partynotvote_7

party respondent would note vote for - seventh ranking

votechoice_partynotvote_8

party respondent would note vote for - eighth ranking

votechoice_partynotvote_9

party respondent would note vote for - ninth ranking

votechoice_partynotvote_text

party respondent would note vote for - text answers

lr_scale

united column of lr_bef and lr_aft values; whether the respondent identifies on the political spectrum

lr_scale_bef

where the respondent identifies on the political spectrum; asked before party identification questions

lr_scale_aft

where the respondent identifies on the political spectrum; asked after party identification questions

religion

religion of respondent

sexuality_selected

sexual identity

sexuality_text

sexual identity; written answers

language_eng

language learned as child and still understand; selected response English

language_fr

language learned as a child and still understand; selected response French

language_abgl

language learned as a child and still understand; specified Aboriginal language

employment

employment status

income

total household income, before taxes, for the year 2018

income_cat

selected household income category

marital

marital status

econ_retro

response to question, 'over the past year, has Canada's economy:'

econ_fed

response to question, 'have the policies of the federal government made Canada's economy...'

econ_self

response to question, have the policies of the federal government made your financial situation...'

Value

The designed dataframe as a 'tbl_df' object under the name decon.

Details

NAs have not been removed. The politically left/right question variables (lr_bef and lr_aft) have also been joined into one column under the name lr_scale. All variables have been converted to factor type using labelled::to_factor and are listed below.

See also

Examples

if (FALSE) {
# call decon dataset
get_decon()

# preview decon
head(decon)
}