Create a dataframe object for a CES survey.

get_ces() creates a dataframe object for a requested Canadian Election Study survey using an associated survey code to call and download the survey dataset. On creation of the data object, prints out the associated citation for use with the requested dataset and a link to the original location of the data file.

get_ces(srvy)

Arguments

srvy
A CES survey code call. See Survey Code Calls below. srvy value must be a character string.

Details

Datasets

Datasets are loaded using .dta, .sav, or .tab file types. See File Types below for a list of included CES datasets and their file type. To quickly convert a dataset's values to factor type use labelled::to_factor() on the dataset.

Survey Code Calls

ces2019_web calls 2019 CES online survey dataset.

ces2019_phone calls 2019 CES phone survey dataset.

ces2015_web calls 2015 CES online survey dataset.

ces2015_phone calls 2015 CES phone survey dataset.

ces2015_combo calls 2015 CES combined (online/phone) dataset.

ces2011 calls 2011 CES survey dataset.

ces2008 calls 2008 CES survey dataset.

ces2004 calls 2004 CES survey dataset.

ces0411 calls 2004-2011 CES survey dataset.

ces0406 calls 2004-2006 CES survey dataset.

ces2000 calls 2000 CES survey dataset.

ces1997 calls 1997 CES survey dataset.

ces1993 calls 1993 CES survey dataset.

ces1988 calls 1988 CES survey dataset.

ces1984 calls 1984 CES survey dataset.

ces1974 calls 1974 CES survey dataset.

ces7480 calls 1974-1980 CES survey dataset.

ces72_jnjl calls 1972 June-July CES survey dataset.

ces72_sep calls 1972 September CES survey dataset.

ces72_nov calls 1972 November CES survey dataset.

ces1968 calls 1968 CES survey dataset.

ces1965 calls 1965 CES survey dataset.

File Types

CES 2019 online survey. Loaded as a .dta file type.

CES 2019 phone survey. Loaded as a .tab file type.

CES 2015 online survey. Loaded as a .dta file type.

CES 2015 phone survey. Loaded as a .dta file type.

CES 2015 combined survey. Loaded as a .dta file type.

CES 2011 survey. Loaded as a .dta file type.

CES 2008 survey. Loaded as a .sav file type.

CES 2004 survey. Loaded as a .sav file type.

CES 2004-2011 surveys. Loaded as a .dta file type.

CES 2004-2006 surveys. Loaded as a .sav file type.

CES 2000 survey. Loaded as a .sav file type.

CES 1997 survey. Loaded as a. sav file type.

CES 1993 survey. Loaded as a .sav file type.

CES 1988 survey. Loaded as a .sav file type.

CES 1974 survey. Loaded as a .sav file type.

CES 1974-1980 surveys. Loaded as a .sav file type.

CES 1972 June-July surveys. Loaded as a .sav file type.

CES 1972 September survey. Loaded as a .sav file type.

CES 1972 November survey. Loaded as a .sav file type.

CES 1968 survey. Loaded as a .sav file type.

CES 1965 survey. Loaded as a .sav file type.

Incorrect/Repeated Code Calls

Incorrect (a non-existent survey code) will stop the function process and return an associated error message. Repeated code calls will load in the raw version of the requested table.

Extra Notes

Due to the naming of the columns in the 1965 and 1968 datasets it is recommended to download the associated codebook for the requested dataset.

Examples

devtools::install_github("hodgettsp/cesR")
#> Downloading GitHub repo hodgettsp/cesR@HEAD
#> 
#>          checking for file 'C:\Users\phodg\AppData\Local\Temp\RtmpMjHD8v\remotes65505b4a1531\hodgettsp-cesR-85f07a4/DESCRIPTION' ...     checking for file 'C:\Users\phodg\AppData\Local\Temp\RtmpMjHD8v\remotes65505b4a1531\hodgettsp-cesR-85f07a4/DESCRIPTION' ...   v  checking for file 'C:\Users\phodg\AppData\Local\Temp\RtmpMjHD8v\remotes65505b4a1531\hodgettsp-cesR-85f07a4/DESCRIPTION' (343ms)
#>       -  preparing 'cesR': (344ms)
#>    checking DESCRIPTION meta-information ...     checking DESCRIPTION meta-information ...   E  checking DESCRIPTION meta-information
#>      Malformed Depends or Suggests or Imports or Enhances field.
#>    Offending entries:
#>      testthat
#>    knitr
#>    Entries must be names of packages optionally followed by '<=' or '>=',
#>    white space, and a valid version number in parentheses.
#>    
#>    See section 'The DESCRIPTION file' in the 'Writing R Extensions'
#>    manual.
#> 
#> Error: Failed to install 'cesR' from GitHub:
#>   System command 'Rcmd.exe' failed, exit status: 1, stdout + stderr:
#> E> * checking for file 'C:\Users\phodg\AppData\Local\Temp\RtmpMjHD8v\remotes65505b4a1531\hodgettsp-cesR-85f07a4/DESCRIPTION' ... OK
#> E> * preparing 'cesR':
#> E> * checking DESCRIPTION meta-information ... ERROR
#> E> Malformed Depends or Suggests or Imports or Enhances field.
#> E> Offending entries:
#> E>   testthat
#> E> knitr
#> E> Entries must be names of packages optionally followed by '<=' or '>=',
#> E> white space, and a valid version number in parentheses.
#> E> 
#> E> See section 'The DESCRIPTION file' in the 'Writing R Extensions'
#> E> manual.

library(cesR)

get_ces("ces2019_web")
#> TO CITE THIS SURVEY FILE:
#> 
#> - Stephenson, Laura B; Harell, Allison; Rubenson, Daniel; Loewen, Peter John, 2020, '2019 Canadian Election Study - Online Survey', https://doi.org/10.7910/DVN/DUS88V, Harvard Dataverse, V1
#> 
#> - Stephenson, Laura, Allison Harrel, Daniel Rubenson and Peter Loewen. Forthcoming. 'Measuring Preferences and Behaviour in the 2019 Canadian Election Study,' Canadian Journal of Political Science.
#> 
#> LINK: https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/DUS88V

ces2019_phone <- labelled::to_factor(ces2019_web)
head(ces2019_web)
#> # A tibble: 6 x 620
#>   cps19_StartDate     cps19_EndDate       cps19_ResponseId         cps19_consent
#>   <dttm>              <dttm>              <chr>                        <dbl+lbl>
#> 1 2019-09-13 08:09:44 2019-09-13 08:36:19 R_1OpYXEFGzHRUpjM 1 [I consent to par~
#> 2 2019-09-13 08:39:09 2019-09-13 08:57:06 R_2qdrL3J618rxYW0 1 [I consent to par~
#> 3 2019-09-13 10:01:19 2019-09-13 10:27:29 R_USWDAPcQEQiMmNb 1 [I consent to par~
#> 4 2019-09-13 10:05:37 2019-09-13 10:50:53 R_3IQaeDXy0tBzEry 1 [I consent to par~
#> 5 2019-09-13 10:05:52 2019-09-13 10:32:53 R_27WeMQ1asip2cMD 1 [I consent to par~
#> 6 2019-09-13 10:10:20 2019-09-13 10:29:45 R_3LiGZcCWJEcWV4P 1 [I consent to par~
#> # ... with 616 more variables: cps19_citizenship <dbl+lbl>,
#> #   cps19_yob <dbl+lbl>, cps19_yob_2001_age <dbl+lbl>, cps19_gender <dbl+lbl>,
#> #   cps19_province <dbl+lbl>, cps19_education <dbl+lbl>,
#> #   cps19_demsat <dbl+lbl>, cps19_imp_iss <chr>, cps19_imp_iss_party <dbl+lbl>,
#> #   cps19_imp_iss_party_7_TEXT <chr>, cps19_imp_loc_iss <chr>,
#> #   cps19_imp_loc_iss_p <dbl+lbl>, cps19_imp_loc_iss_p_7_TEXT <chr>,
#> #   cps19_interest_gen_1 <dbl>, cps19_interest_elxn_1 <dbl>, ...