Creates a non-exhaustive dataframe of 21 variables with renamed columns.

get_decon() creates a non-exhaustive dataset under the name decon consisting of 21 variables with renamed columns from the demographics, ideology, and economy sections of the 2019 CES online survey.

get_decon()

Details

NAs have not been removed. The politically left/right question variables (lr_bef and lr_aft) have also been joined in one column under the name lr. All variables have been converted to factor type using labelled::to_factor and are listed below.

decon Variables

citizenship

Canadian citizenship status

yob

year of birth

gender

identified gender of the respondent

province_territory

Province or Territory of current residence

education

highest level of education completed

lr

united column of lr_bef and lr_aft values; whether the respondent identifies on the political spectrum

lr_bef

where the respodent identifies on the political spectrum; asked before party identification questions

lr_aft

where the respondent identifies on the political spectrum; asked after party identification questions

religion

religion of respondent

sexuality_selected

sexual identity

sexuality_text

sexual identity; written answers

language_eng

language learned as child and still understand; selected response English

language_fr

language learned as a child and still understand; selected response French

language_abgl

langauge learned as a child and still understand; specified Aboriginal language

employment

employment status

income

total household income, before taxes, for the year 2018

income_cat

selected household income category

marital

marital status

econ_retro

response to question, 'over the past year, has Canada's economy:'

econ_fed

response to question, 'have the policies of the federal government made Canada's economy...'

econ_self

response to question, have the policies of the federal government made your financial situation...'

See also

Examples

devtools::install_github("hodgettsp/cesR")

#> Downloading GitHub repo hodgettsp/cesR@HEAD

#> 
#>          checking for file 'C:\Users\phodg\AppData\Local\Temp\Rtmpag3bIp\remotes5d885adf216e\hodgettsp-cesR-e518ad2/DESCRIPTION' ...  v  checking for file 'C:\Users\phodg\AppData\Local\Temp\Rtmpag3bIp\remotes5d885adf216e\hodgettsp-cesR-e518ad2/DESCRIPTION'
#>       -  preparing 'cesR':
#>    checking DESCRIPTION meta-information ...     checking DESCRIPTION meta-information ...   v  checking DESCRIPTION meta-information
#>       -  checking for LF line-endings in source and make files and shell scripts
#>   -  checking for empty or unneeded directories
#>       -  building 'cesR_0.1.0.tar.gz'
#>      
#> 

#> Error: Failed to install 'cesR' from GitHub:
#>   (converted from warning) package 'cesR' is in use and will not be installed


library(cesR)

get_decon()

#> TO CITE THIS SURVEY FILE: Stephenson, Laura B; Harell, Allison; Rubenson, Daniel; Loewen, Peter John, 2020, '2019 Canadian Election Study - Online Survey',
#>            https://doi.org/10.7910/DVN/DUS88V, Harvard Dataverse, V1
#> LINK: https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/DUS88V

head(decon)

#> # A tibble: 6 x 22
#>   ces_code citizenship yob   gender province_territ~ education lr    lr_bef
#>   <chr>    <fct>       <fct> <fct>  <fct>            <fct>     <chr> <chr> 
#> 1 ces2019~ Canadian c~ 1989  A wom~ Quebec           Master's~ 2     NA    
#> 2 ces2019~ Canadian c~ 1998  A wom~ Quebec           Master's~ 2     NA    
#> 3 ces2019~ Canadian c~ 2000  A wom~ Ontario          Some uni~ 4     4     
#> 4 ces2019~ Canadian c~ 1998  A man  Ontario          Some uni~ 7     7     
#> 5 ces2019~ Canadian c~ 2000  A wom~ Ontario          Complete~ 4     4     
#> 6 ces2019~ Canadian c~ 1999  A wom~ Ontario          Some uni~ 4     4     
#> # ... with 14 more variables: lr_aft <chr>, religion <fct>,
#> #   sexuality_selected <fct>, sexuality_text <chr>, language_eng <fct>,
#> #   language_fr <fct>, language_abgl <fct>, employment <fct>, income <dbl>,
#> #   income_cat <fct>, marital <fct>, econ_retro <fct>, econ_fed <fct>,
#> #   econ_self <fct>